+++ index.js
@@ -4,10 +4,12 @@ const ipv4 = require('./src/ipv4/ipv4')
 const validateJsonSchema = require('./src/validateJsonSchema/validateJsonSchema')
 const verifyJWT = require('./src/signAndVerifyJWT/verifyJWT')
 const signJWT = require('./src/signAndVerifyJWT/signJWT')
+const parseTypes = require('./src/parseTypes/parseTypes')
 
 module.exports = {
   ipv4: ipv4,
   validateJsonSchema: validateJsonSchema,
   verifyJWT: verifyJWT,
-  signJWT: signJWT
+  signJWT: signJWT,
+  parseTypes: parseTypes
 }